+++ src/core/managers/internal/LedgerConnectStateManager/LedgerConnectStateManager.ts
@@ -140,6 +140,10 @@ export class LedgerConnectStateManager extends SidePanelBaseManager<
       LedgerConnectEventsEnum.CLOSE_LEDGER_CONNECT_PANEL,
       onCancel
     );
+    this.eventBus.subscribe(
+      LedgerConnectEventsEnum.UI_DISCONNECTED,
+      this.destroy.bind(this)
+    );
   }
 
   public unsubscribeFromProviderInit(
@@ -154,6 +158,10 @@ export class LedgerConnectStateManager extends SidePanelBaseManager<
     this.eventBus.unsubscribe(
       LedgerConnectEventsEnum.CLOSE_LEDGER_CONNECT_PANEL,
       onCancel
+    );
+    this.eventBus.unsubscribe(
+      LedgerConnectEventsEnum.UI_DISCONNECTED,
+      this.destroy.bind(this)
     );
   }