+++ src/core/providers/strategies/CrossWindowProviderStrategy/CrossWindowProviderStrategy.ts
@@ -14,15 +14,15 @@ import { ProviderErrorsEnum } from 'types/provider.types';
 import { getPendingTransactionsHandlers } from '../helpers/getPendingTransactionsHandlers';
 import { signMessage } from '../helpers/signMessage/signMessage';
 import { guardTransactions } from '../helpers/signTransactions/helpers/guardTransactions/guardTransactions';
+import { BaseProviderStrategy } from '../BaseProviderStrategy/BaseProviderStrategy';
 
 type CrossWindowProviderProps = {
   address?: string;
   walletAddress?: string;
 };
 
-export class CrossWindowProviderStrategy {
+export class CrossWindowProviderStrategy extends BaseProviderStrategy {
   private provider: CrossWindowProvider | null = null;
-  private address: string;
   private walletAddress?: string;
   private _signTransactions:
     | ((transactions: Transaction[]) => Promise<Transaction[]>)
@@ -31,7 +31,7 @@ export class CrossWindowProviderStrategy {
     null;
 
   constructor(config?: CrossWindowProviderProps) {
-    this.address = config?.address || '';
+    super(config?.address);
     this.walletAddress = config?.walletAddress;
   }
 
@@ -47,6 +47,7 @@ export class CrossWindowProviderStrategy {
     // Bind in order to break reference
     this._signTransactions = this.provider.signTransactions.bind(this.provider);
     this._signMessage = this.provider.signMessage.bind(this.provider);
+    this._login = this.provider.login.bind(this.provider);
 
     this.provider.setWalletUrl(this.walletAddress || network.walletAddress);
     this.provider.setAddress(this.address);
@@ -56,32 +57,23 @@ export class CrossWindowProviderStrategy {
     return this.buildProvider();
   };
 
+  protected override cancelAction() {
+    return this.provider?.cancelAction?.bind(this.provider)();
+  }
+
   private buildProvider = () => {
     if (!this.provider) {
       throw new Error(ProviderErrorsEnum.notInitialized);
     }
 
     const provider = this.provider as unknown as IProvider;
-
     provider.setAccount({ address: this.address });
     provider.signTransactions = this.signTransactions;
     provider.signMessage = this.signMessage;
+    provider.login = this.login;
+    provider.cancelLogin = this.cancelLogin;
 
     return provider;
-  };
-
-  private initialize = () => {
-    if (this.address) {
-      return;
-    }
-
-    const address = getAddress();
-
-    if (!address) {
-      return;
-    }
-
-    this.address = address;
   };
 
   private signTransactions = async (transactions: Transaction[]) => {
@@ -145,7 +137,6 @@ export class CrossWindowProviderStrategy {
 
     return signedMessage;
   };
-
   private setPopupConsent = () => {
     const crossWindowDappConfig = crossWindowConfigSelector(getState());